requirements.txt
selenium==4.9.1
python-telegram-bot~=20.02
schedule